+++ src/main/java/com/wanted/naeil/domain/admin/entity/enums/ApprovalStatus.java
@@ -1,16 +1,14 @@
-package com.wanted.naeil.domain.live.entity;
+package com.wanted.naeil.domain.admin.entity.enums;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum LiveLectureStatus {
+public enum ApprovalStatus {
     PENDING("승인 대기"),
     APPROVED("승인 완료"),
-    REJECTED("반려됨"),
-    IN_PROGRESS("방송 중"),
-    ENDED("종료됨");
+    REJECTED("반려");
 
     private final String description;
 }
