+++ src/main/java/com/wanted/naeil/domain/admin/repository/AdminApprovalRepository.java
@@ -1,7 +1,7 @@
 package com.wanted.naeil.domain.admin.repository;
 
-import com.wanted.naeil.domain.admin.entity.ApprovalRequestType;
-import com.wanted.naeil.domain.admin.entity.ApprovalStatus;
+import com.wanted.naeil.domain.admin.entity.enums.ApprovalRequestType;
+import com.wanted.naeil.domain.admin.entity.enums.ApprovalStatus;
 import com.wanted.naeil.domain.admin.entity.AdminApproval;
 import org.springframework.data.jpa.repository.JpaRepository;
