+++ src/main/java/com/wanted/naeil/domain/admin/service/AdminApprovalService.java
@@ -2,15 +2,15 @@ package com.wanted.naeil.domain.admin.service;
 
 import com.wanted.naeil.domain.admin.dto.response.ApprovalResponse;
 import com.wanted.naeil.domain.admin.entity.AdminApproval;
-import com.wanted.naeil.domain.admin.entity.ApprovalRequestType;
-import com.wanted.naeil.domain.admin.entity.ApprovalStatus;
+import com.wanted.naeil.domain.admin.entity.enums.ApprovalRequestType;
+import com.wanted.naeil.domain.admin.entity.enums.ApprovalStatus;
 import com.wanted.naeil.domain.admin.repository.AdminApprovalRepository;
 import com.wanted.naeil.domain.course.entity.Course;
 import com.wanted.naeil.domain.live.entity.LiveLecture;
-import com.wanted.naeil.domain.live.entity.LiveLectureStatus;
+import com.wanted.naeil.domain.live.entity.enums.LiveLectureStatus;
 import com.wanted.naeil.domain.live.repository.LiveLectureRepository;
 import com.wanted.naeil.domain.user.entity.InstructorApplications;
-import com.wanted.naeil.domain.user.entity.Role;
+import com.wanted.naeil.domain.user.entity.enums.Role;
 import com.wanted.naeil.domain.user.entity.User;
 import com.wanted.naeil.domain.user.repository.InsturctorApplicationRepository;
 import lombok.RequiredArgsConstructor;
