+++ src/main/java/com/wanted/naeil/domain/community/entity/enums/LikeTargetType.java
@@ -1,12 +1,13 @@
-package com.wanted.naeil.domain.community.entity;
+package com.wanted.naeil.domain.community.entity.enums;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum PostCategory {
-    FREE("자유"), QNA("QnA");
+public enum LikeTargetType {
+    POST("게시글"),
+    COURSE("코스");
 
     private final String description;
 }
