+++ src/main/java/com/wanted/naeil/domain/community/repository/PostRepository.java
@@ -1,7 +1,7 @@
 package com.wanted.naeil.domain.community.repository;
 
 import com.wanted.naeil.domain.community.entity.Post;
-import com.wanted.naeil.domain.community.entity.PostCategory;
+import com.wanted.naeil.domain.community.entity.enums.PostCategory;
 import com.wanted.naeil.domain.course.entity.Course;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
