+++ src/main/java/com/wanted/naeil/domain/course/entity/Section.java
@@ -1,5 +1,6 @@
 package com.wanted.naeil.domain.course.entity;
 
+import com.wanted.naeil.domain.course.entity.enums.SectionStatus;
 import com.wanted.naeil.global.common.entity.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.Builder;
