+++ src/main/java/com/wanted/naeil/domain/course/entity/enums/CourseStatus.java
@@ -1,13 +1,14 @@
-package com.wanted.naeil.domain.admin.entity;
+package com.wanted.naeil.domain.course.entity.enums;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum ApprovalStatus {
+public enum CourseStatus {
     PENDING("승인 대기"),
-    APPROVED("승인 완료"),
+    ACTIVE("활성화"),
+    INACTIVE("비활성화"),
     REJECTED("반려");
 
     private final String description;
