+++ src/main/java/com/wanted/naeil/domain/learning/entity/LearningProgress.java
@@ -1,6 +1,7 @@
 package com.wanted.naeil.domain.learning.entity;
 
 import com.wanted.naeil.domain.course.entity.Section;
+import com.wanted.naeil.domain.learning.entity.enums.ProgressStatus;
 import com.wanted.naeil.domain.user.entity.User;
 import com.wanted.naeil.global.common.entity.BaseTimeEntity;
 import jakarta.persistence.*;
