+++ src/main/java/com/wanted/naeil/domain/learning/entity/enums/EnrollmentStatus.java
@@ -1,12 +1,13 @@
-package com.wanted.naeil.domain.learning.entity;
+package com.wanted.naeil.domain.learning.entity.enums;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum ProgressStatus {
-    NOT_STARTED("수강 전"),
+public enum EnrollmentStatus {
+
+    BEFORE_START("수강 전"),
     IN_PROGRESS("수강 중"),
     COMPLETED("수강 완료");
