+++ src/main/java/com/wanted/naeil/domain/live/entity/enums/LiveLectureStatus.java
@@ -1,16 +1,16 @@
-package com.wanted.naeil.domain.settlement.entity;
+package com.wanted.naeil.domain.live.entity.enums;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum SettlementStatus {
-
+public enum LiveLectureStatus {
     PENDING("승인 대기"),
     APPROVED("승인 완료"),
-    REJECTED("반려"),
-    CANCELED("강사 취소");
+    REJECTED("반려됨"),
+    IN_PROGRESS("방송 중"),
+    ENDED("종료됨");
 
     private final String description;
 }
