+++ src/main/java/com/wanted/naeil/domain/payment/entity/PaymentItem.java
@@ -1,8 +1,8 @@
 package com.wanted.naeil.domain.payment.entity;
 
 import com.wanted.naeil.domain.course.entity.Course;
+import com.wanted.naeil.domain.payment.entity.enums.PaymentItemType;
 import jakarta.persistence.*;
-import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
