+++ src/main/java/com/wanted/naeil/domain/settlement/entity/Settlement.java
@@ -1,9 +1,8 @@
 package com.wanted.naeil.domain.settlement.entity;
 
+import com.wanted.naeil.domain.settlement.entity.enums.SettlementStatus;
 import com.wanted.naeil.domain.user.entity.User;
-import com.wanted.naeil.global.common.entity.BaseTimeEntity;
 import jakarta.persistence.*;
-import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
