+++ src/main/java/com/wanted/naeil/domain/settlement/entity/enums/SettlementStatus.java
@@ -1,14 +1,16 @@
-package com.wanted.naeil.domain.user.entity;
+package com.wanted.naeil.domain.settlement.entity.enums;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum ApplicationStatus {
+public enum SettlementStatus {
+
     PENDING("승인 대기"),
     APPROVED("승인 완료"),
-    REJECTED("반려");
+    REJECTED("반려"),
+    CANCELED("강사 취소");
 
     private final String description;
 }
