+++ src/main/java/com/wanted/naeil/domain/user/entity/enums/ApplicationStatus.java
@@ -1,14 +1,13 @@
-package com.wanted.naeil.domain.course.entity;
+package com.wanted.naeil.domain.user.entity.enums;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum CourseStatus {
+public enum ApplicationStatus {
     PENDING("승인 대기"),
-    ACTIVE("활성화"),
-    INACTIVE("비활성화"),
+    APPROVED("승인 완료"),
     REJECTED("반려");
 
     private final String description;
