+++ src/main/java/com/wanted/naeil/global/common/controller/IndexController.java
@@ -1,7 +1,7 @@
 package com.wanted.naeil.global.common.controller;
 
 import com.wanted.naeil.domain.auth.model.dto.AuthDetails;
-import com.wanted.naeil.domain.user.entity.Role;
+import com.wanted.naeil.domain.user.entity.enums.Role;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
